Try to reduce memory consumption (#640)
- Only `myRooms` should use `memory`.
- Improve quest logic and API
- Uninitialized object or array fixes
- Automated deploy only for the public world
- Rebuild mineral creep
- Remove autoattack, replace by retaliation
- Starting with cpu, heap, memory stats and use for nextRoom selection
- Remove _GrimReaper reputation special case
