+++ doc/NPC.md
@@ -1,7 +1,7 @@
 # TooAngel NPC
 
 Welcome stranger. Nice to see you here.
-Let me introduce myself, I'm the TooAngel NPC a [community driven](https://www.worlddriven.org) account. If you like to contact with makers head over to [discord](https://discord.com/channels/860665589738635336/931237079315251280).
+Let me introduce myself, I'm the TooAngel NPC a [community driven](https://www.worlddriven.org) account. If you like to get in contact with makers head over to [discord](https://discord.com/channels/860665589738635336/931237079315251280).
 
 ## Reputation
 
@@ -9,7 +9,7 @@ From here on you can decide how to interact with the TooAngel NPC, based on your
 
 I don't like:
 
-- Send creeps into my rooms
+- Hostile creeps into my rooms
 - Destroying structures of even spawns
 - Sending nukes
 
@@ -34,23 +34,10 @@ Based on the level of reputation certain actions will be performed, like:
 
 ## Basic behavior
 
-- I want to make sure that you are prepared for the dangers out there. On the first meetings I'll make sure that your room defense is up for a minimal standard.
-
-See it as a:
-
-"It is most gratifying that your enthusiasm for our planet continues unabated.
-As a token of our appreciation, we hope you will enjoy the two thermonuclear
-missiles we've just sent to converge with your craft. To ensure ongoing quality
-of service, your death may be monitored for training purposes. Thank you."
-(The Hitchhiker's Guide to the Galaxy)
-
-
-- There is a radius around my rooms which I see as mine, all interactions there are considered hostile and I'll react appropiate.
-
-
-- I like my rooms, if attacked or destroyed, I'll try my best to keep them, even after loosing the last spawn.
+- I keep a safety radius of 1 linear distance around the rooms and attack other controlled or reserved rooms
+- Hostile actions are retaliated
+- I like my rooms, if attacked or destroyed, I'll try the best to keep them, even after loosing the last spawn.
 
 ## API
 
-- Sending resources via terminal will increase the credit value of the resources to the `reputation`
-- Segements 1, 2 are public and show the Top and Bottom players based on their `reputation` 
+For communication head over to the [API](API.md)
