+++ package.json
@@ -4,9 +4,8 @@
   "description": "",
   "main": "src/main.js",
   "screeps_bot": true,
-  "dependencies": {},
   "devDependencies": {
-    "chai": "4.3.4",
+    "chai": "4.3.6",
     "eslint-config-google": "0.14.0",
     "eslint-plugin-screeps": "2.1.0",
     "grunt": "1.4.1",
@@ -18,7 +17,7 @@
     "grunt-mocha-test": "0.13.3",
     "grunt-screeps": "1.5.0",
     "grunt-sync": "0.8.2",
-    "mocha": "9.1.3",
+    "mocha": "9.2.0",
     "request": "2.88.2",
     "request-promise-native": "1.0.9",
     "rimraf": "3.0.2",
