+++ src/brain_squadmanager.js
@@ -1,5 +1,7 @@
 'use strict';
 
+const {debugLog} = require('./logging');
+
 brain.isFriend = function(name) {
   if (!Memory.players) {
     Memory.players = {};
@@ -25,7 +27,7 @@ brain.isFriend = function(name) {
 
 brain.handleSquadmanager = function() {
   if (Object.keys(Memory.squads).length > 0) {
-    brain.debugLog('brain', 'brain.handleSquadmanager squads: ${Object.keys(Memory.squads).length}');
+    debugLog('brain', 'brain.handleSquadmanager squads: ${Object.keys(Memory.squads).length}');
   }
   for (const squadIndex of Object.keys(Memory.squads)) {
     const squad = Memory.squads[squadIndex];
