+++ src/diplomacy.js
@@ -1,3 +1,6 @@
+const {debugLog} = require('./logging');
+const {getMyRoomWithinRange} = require('./helper_findMyRooms');
+
 /**
  * checkPlayers
  *
@@ -12,9 +15,6 @@ function checkPlayers() {
   }
   Memory.players = Memory.players || {};
   for (const playerName of Object.keys(Memory.players)) {
-    if (playerName === '_GrimReaper') {
-      player.reputation = Math.abs(player.reputation);
-    }
     const player = Memory.players[playerName];
     player.reputation = player.reputation || 0;
     for (const roomName of Object.keys(player.rooms || {})) {
@@ -48,6 +48,95 @@ function checkPlayers() {
 
 module.exports.checkPlayers = checkPlayers;
 
+const findRoomPairs = function(player) {
+  for (const roomName of Object.keys(player.rooms).sort((a, b) => 0.5 - Math.random())) {
+    debugLog('diplomacy', `findRoomPairs: room ${roomName} data: ${JSON.stringify(global.data.rooms[roomName])}`);
+    const minRCL = ((global.data.rooms[roomName] || {}).controller || {}).level || 8;
+    const range = 7;
+    const myRoomName = getMyRoomWithinRange(roomName, range, minRCL);
+    if (myRoomName) {
+      return {
+        myRoomName: myRoomName,
+        theirRoomName: roomName,
+      };
+    }
+  }
+  return false;
+};
+
+/**
+ * handleRetaliation
+ *
+ * @param {object} player
+ * @return {void}
+ */
+function handleRetaliation(player) {
+  if (!player.lastAttacked) {
+    player.lastAttacked = Game.time + config.autoattack.timeBetweenAttacks;
+  }
+  if (Game.time < player.lastAttacked + config.autoattack.timeBetweenAttacks) {
+    debugLog('diplomacy', `Too early to attack`);
+    return false;
+  }
+  const actions = [
+    {
+      name: 'simpleAttack',
+      value: -1 * 1500,
+      level: 0,
+      execute: (roomPair) => {
+        const origin = Game.rooms[roomPair.myRoomName];
+        origin.checkRoleToSpawn('autoattackmelee', 1, undefined, roomPair.theirRoomName);
+      },
+    },
+    {
+      name: 'squad',
+      value: -4 * 1500,
+      level: 1,
+      execute: (roomPair) => {
+        brain.startSquad(roomPair.myRoomName, roomPair.theirRoomName);
+      },
+    },
+    {
+      name: 'attack42',
+      value: -6 * 1500,
+      level: 2,
+      execute: (roomPair) => {
+        brain.startMeleeSquad(roomPair.myRoomName, roomPair.theirRoomName);
+      },
+    },
+  ];
+  const possibleActions = actions.filter((action) => {
+    if (action.value < player.reputation) {
+      return false;
+    }
+    if (action.level < player.level) {
+      return false;
+    }
+    return true;
+  });
+  if (possibleActions.length === 0) {
+    return false;
+  }
+  debugLog('diplomacy', `handleRetaliation player: ${JSON.stringify(player)} possibleActions: ${JSON.stringify(possibleActions)}`);
+  const roomPair = findRoomPairs(player);
+  if (!roomPair) {
+    debugLog('diplomacy', `handleRetaliation: Can not find a fitting room pair`);
+    return;
+  }
+  possibleActions.sort((a, b) => 0.5 - Math.random());
+  debugLog('diplomacy', `Running attach roomPair: ${JSON.stringify(roomPair)} action: ${JSON.stringify(possibleActions[0])}`);
+  player.lastAttacked = Game.time;
+  if (config.autoattack.notify) {
+    Game.notify(Game.time + ' ' + this.name + ' Queuing retaliation');
+  }
+  player.counter++;
+  if (player.counter > 10) {
+    player.level += 1;
+    player.counter = 0;
+  }
+  possibleActions[0].execute(roomPair);
+}
+
 /**
  * addToReputation
  *
@@ -61,17 +150,25 @@ function addToReputation(name, value) {
   if (name === 'Invader') {
     return;
   }
-
   value = value || 0;
   Memory.players = Memory.players || {};
 
-  initPlayer(name);
+  const player = initPlayer(name);
+  debugLog('diplomacy', `addToReputation name: ${name} value: ${value} player: ${JSON.stringify(player)}`);
 
-  if (!Memory.players[name].reputation) {
-    Memory.players[name].reputation = 0;
+  if (!player.reputation) {
+    player.reputation = 0;
   }
 
-  Memory.players[name].reputation += value;
+  player.reputation += value;
+
+  if (value < 0) {
+    try {
+      handleRetaliation(player);
+    } catch (e) {
+      console.log(e);
+    }
+  }
 }
 
 module.exports.addToReputation = addToReputation;
@@ -80,6 +177,7 @@ module.exports.addToReputation = addToReputation;
  * initPlayer
  *
  * @param {string} name
+ * @return {object}
  */
 function initPlayer(name) {
   if (!Memory.players[name]) {
@@ -91,5 +189,19 @@ function initPlayer(name) {
       reputation: 0,
     };
   }
+  return Memory.players[name];
 }
 module.exports.initPlayer = initPlayer;
+
+function addRoomToPlayer(player, room) {
+  if (!player.rooms) {
+    player.rooms = {};
+  }
+  if (!player.rooms[room.name]) {
+    player.rooms[room.name] = {
+      visited: Game.time,
+    };
+  }
+}
+
+module.exports.addRoomToPlayer = addRoomToPlayer;
