+++ src/logging.js
@@ -1,8 +1,10 @@
-brain.debugLog = function(type, ...messages) {
+function debugLog(type, ...messages) {
   if (config.debug[type]) {
     console.log(`${Game.time} ${messages.join(' ')}`);
   }
-};
+}
+
+module.exports.debugLog = debugLog;
 
 Room.prototype.log = function(...messages) {
   console.log(`${Game.time} ${this.name.rpad(' ', 27)} ${messages.join(' ')}`);
