+++ src/prototype_room_mineral.js
@@ -115,9 +115,7 @@ Room.prototype.reactions = function() {
 
   if (this.getResourceAmountWithNextTiers(this.memory.reaction.result.result) > config.mineral.minAmount &&
     (this.terminal.store[this.memory.reaction.result.result] > config.mineral.minAmount)) {
-    if (config.debug.mineral) {
-      this.log('Done with reaction:' + this.memory.reaction.result.result);
-    }
+    this.debugLog('mineral', 'Done with reaction:' + this.memory.reaction.result.result);
     delete this.memory.reaction;
   }
 };
