+++ src/require.js
@@ -5,7 +5,6 @@ require('./brain_main');
 require('./brain_memory');
 require('./brain_memory_market');
 require('./brain_nextroom');
-require('./brain_quests');
 require('./brain_squadmanager');
 require('./brain_stats');
 require('./prototype_creep');
@@ -18,7 +17,6 @@ require('./prototype_creep_mineral');
 require('./prototype_creep_routing');
 require('./prototype_roomPosition_structures');
 require('./prototype_room');
-require('./prototype_room_attack');
 require('./prototype_room_basebuilder');
 require('./prototype_room_controller');
 require('./prototype_room_creepbuilder');
